pyproject.toml
[project]
name = "katcha"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.20.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0.0",
]
